[UPDATE]플레이어 , 폭탄 스크립트 수정
플레이어가 폭탄에 닿으면 획득
