--- Assets/Scripts/Items/Boom.cs
@@ -1,43 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class Boom : MonoBehaviour
-{
-    public float explosionForce = 10f; 
-    public float explosionRadius = 5f; 
-
-    private void OnCollisionEnter(Collision collision)
-    {
-        if(collision.gameObject.CompareTag("Wall"))
-        {
-            Explode();
-        }
-    }
-
-    void Explode()
-    {
-        Collider[] colliders = Physics.OverlapSphere(transform.position, explosionRadius);
-
-        foreach (var collider in colliders)
-        {
-            Rigidbody rb = collider.GetComponent<Rigidbody>();
-            if (rb != null)
-            {
-                rb.AddExplosionForce(explosionForce, transform.position, explosionRadius);
-            }
-
-            DestroyWall(collider.gameObject);
-        }
-
-      
-        Destroy(gameObject);
-    }
-
-    void DestroyWall(GameObject wall)
-    {
-        
-        Destroy(wall);
-    }
-}
-
