--- Assets/Scripts/Items/Boom.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 84cc6a09c22795949aea2083272ab9e2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
